refactor request log skeleton to separate file
