+++ src/components/skeletons/index.ts
@@ -1,3 +1,4 @@
 export { default as Table } from "./table"
 export { default as EventLogs } from "./event-logs"
+export { default as RequestLogs } from "./request-logs"
 export { default as PolicyFields } from "./policy-fields"
